Make the parse_func_t take a const string ref.
This can improve performance by quite a bit, so I'm doing it. I assume
anyone with a brain can figure out that they should make a copy of the
string if they want to keep it around. A const string ref isn't much
use except as a parameter anyway. Don't try to correct me on that.