Move some member funcs into state_t.
Also moves func_ into state_t. This is mostly because the state and
the parser are somewhat independent (the parser just works on the
state, so I see it).

Basically, this is all in the name of mostly pointless performance. In
the end, this cuts 5 seconds off the parser when it's used to parse a
small document 2^20 times. Not really important, but it's a feel-good
thing I can do.