Fix destructor issue in object_pool_t
Because the pool depends on std::vector by default, the objects in the
vector would be destroyed as usual. This presents some odd problems for
objects that have already been destroyed by collect(), so the pooled
objects are now dummy objects of the same size as the objects held by
the pool. This could probably go wrong in certain situations, but if
that happens, it's really probably time for something other than a
generic object pool.