Change string internals a bit
Now uses as much of the length and capacity variables as it can. This
should allow it to avoid additional allocations for shorter strings.
Downside is that it does make the code slightly more complicated, but
not as much as I thought it would. Did have to resort to a sort of hack
to get a fixed size for the short data array, though.