Remove use of index set to store unused indices.
This is probably slightly slower, but the storage used should be about
the same or less in pools. This does have the upside of filling up
earlier holes in the pool, though this is also doable given a sorted
and ordered index container although it's likely to be a tiny bit
costly to remove indices from the container.