Alias string_t::size_type to int.
Although using size_t makes some sense, it's rarely actually useful.
Basically, the chance of anyone needing a string consisting of more
than 2^31 bytes is sort of really low, at least as far as I'm
concerned, and having to deal with the signed/unsigned difference when
iterating using indices is a pain. So, rather than waste time with
size_t, I'm just going to switch it to int.