Add a string_t::reserve_for_growth function.
Because it's not really convenient to reallocate a string's underlying
buffer every time a char is appended and the new reserve() behavior
allocates only the requested amount if it's larger than the current
capacity, I've added a reserve_for_growth member function. It's private
for the time being since it's not very flexible right now.

reserve_for_growth is only called when the buffer is resized by an
operation like append, insert, or push_back. It currently grows the
buffer by 40% only if the consumed size of the buffer is nearly 85% --
if the needed capacity is greater than the existing capacity and
greater than the speculative capacity (the 40% growth), then it will
reserve the needed capacity and no more than that.

Calling resize() will not attempt to reserve for growth, so it's
entirely possible to control growth yourself by reserving a different
capacity and resizing/modifying the string without using the append or
insert functions.