align: Add align.hh
Has one function with two ways of calling it:

- align(size_t num, size_t alignment) -> size_t
- align<size_t alignment>(size_t num) -> size_t

Both are constexpr. The latter is specialized for alignment 1 to simply
return num. Returns num aligned to the given alignment. Doesn't check
that alignment is a power of two, though will probably produce very
bad results if it isn't. Use your brain.