Add io.hh
Covers basic snow::io namespace.

Borrowed mainly from scolex's io_ops.hh -- provides a basic procedural
interface to streams and stream-likes. Does not include the string
class specific functions nor the eof() for types that don't provide
an eof(). That may be re-added later, but for now it's omitted because
it's hacky.

Neither read/write/seek/tell depend on the existence of eof, so it
shouldn't affect anything that doesn't expect eof to exist.